Add Register(Admin&Company)  && Create usre -Get All User

+++ api/TalentFlow.Application/Models/Identity/JwtSettings.cs
@@ -5,7 +5,8 @@
         public string Key { get; set; } = string.Empty;
         public string Issuer { get; set; } = string.Empty;
         public string Audience { get; set; } = string.Empty;
-        public double DurationInMinutes { get; set; }
+        public int DurationInMinutes { get; set; }
         public int RefreshTokenDurationInDays { get; set; }
     }
 }
+
